+++ react/src/Home.tsx
@@ -12,9 +12,11 @@ function Home() {
     const [user, setUser] = useState<user | null>(null);
     const [projects, setProjects] = useState<project[]>([]);
     const [selectedProject, setSelectedProject] = useState<project | null>(null);
+    const [isWidgetStaged, setIsWidgetStaged] = useState<boolean>(false);
     const navigate = useNavigate();
     const gridRef = useRef<any>(null);
 
+    // Handles selecting project from dropdown
     function HandleProjectChange(id : number) : void {
         const proj : project | undefined = projects.find((p : project) => p.id == id);
         if (proj) {
@@ -22,23 +24,30 @@ function Home() {
         }
     }
 
-    // Calls AddWidget in GridSystem component
+    // Stages a widget in GridSystem component
     function StageWidget() : void {
-        if (gridRef.current) {
+        if (gridRef.current && !isWidgetStaged) {
             gridRef.current.StageWidget();
         }
     }
+    // Cancels the staged widget in GridSystem component
+    function UnstageWidget() : void {
+        if (gridRef.current && isWidgetStaged) {
+            gridRef.current.UnstageWidget();
+        }
+    }
 
+    // Initial mount
     useEffect(() => {
         // Fetches the currently logged in user
-        async function FetchUser() {
+        async function FetchUser() : Promise<void> {
             const res : response = await window.api.GetCurrentUser();
             if (res.data) {
                 setUser(res.data);
             }
             // No user logged in
             else {
-                toast.error("Whoops! Looks like your login has expired");
+                toast.error("No user logged in");
                 setTimeout(() => navigate('/'), 3000); // 3s delay
             }
         }
@@ -48,7 +57,7 @@ function Home() {
 
     useEffect(() => {
         // Fetches the user's projects
-        async function FetchProjects(username : string) {
+        async function FetchProjects(username : string) : Promise<void> {
             const res : response = await window.api.GetProjects(username);
             if (res.err) {
                 toast.error("Error retrieving projects");
@@ -67,20 +76,29 @@ function Home() {
         <ToastContainer limit={1} transition={Flip} />
         <div className="home-container">
             <div className="sidebar">
+                {/* User settings button */}
                 <button className="user-info d-flex align-items-center">
-                    <img className="mt-2 ms-2" src="/default_pfp.png"></img>
-                    {user ? <h5 className="ms-2">{user.f_name} {user.l_name}</h5> : <h5>Loading...</h5>}
+                    <img className="mt-2 ms-2" src="/default_pfp.png"/>
+                    {user ? <h5 className="ms-2">{user.f_name} {user.l_name}</h5> : 
+                    <h5>Loading...</h5>}
                 </button>
             </div>
             <div className="navbar">
+                {/* Select project dropdown */}
                 <div className="dropdown">
-                    <button type="button" className="btn btn-primary dropdown-toggle" data-bs-toggle="dropdown">
+                    <button 
+                        type="button" 
+                        className="btn btn-primary dropdown-toggle" 
+                        data-bs-toggle="dropdown"
+                    >
                         {selectedProject ? selectedProject.name : "Select Palette"}
                     </button>
                     <ul className="dropdown-menu">
-                        {projects.length ? (projects.map((proj : project, index : number) => (
+                        {projects.length ? (projects.map((proj : project) => (
                             <li key={proj.id}>
-                                <a className="dropdown-item" href="#" onClick={() => HandleProjectChange(proj.id)}>
+                                <a className="dropdown-item" href="#" onClick={
+                                    () => HandleProjectChange(proj.id)
+                                }>
                                     {proj.name}
                                 </a>
                             </li>
@@ -90,11 +108,23 @@ function Home() {
                                 </h6>
                               </li>)}
                     </ul>
-                    <button onClick={()=>StageWidget()}>New widget</button>
+                    {/* Create widget or cancel staged widget */}
+                    {selectedProject && gridRef.current ? (
+                        <button onClick={
+                            () => isWidgetStaged ? UnstageWidget() : StageWidget()
+                        }>{isWidgetStaged ? 'Cancel' : 'New widget'}</button>
+                    ) : (
+                        ''
+                    )}
                 </div>
             </div>
             <div className="content">
-                <GridSystem project={selectedProject} ref={gridRef} />
+                {/* Grid displaying all project widgets */}
+                <GridSystem 
+                    project={selectedProject} 
+                    setIsWidgetStaged={setIsWidgetStaged}
+                    ref={gridRef} 
+                />
             </div>
         </div>
         </>